rename TraitList folder
