+++ packages/api/schema.gql
@@ -49,7 +49,7 @@ type Critter {
   ownerId: ID!
   species: Species!
   speciesId: ID!
-  traitList: TraitList!
+  traitList: Variant!
   traitListId: ID!
   traitValues: [CritterTraitValue!]!
 }
@@ -126,7 +126,7 @@ type EnumValueSetting {
   enumValue: EnumValue!
   enumValueId: ID!
   id: ID!
-  traitList: [TraitList!]!
+  traitList: [Variant!]!
   traitListId: ID!
 }
 
@@ -308,7 +308,7 @@ type Species {
 
   """Name of the species"""
   name: String!
-  traitLists: [TraitList!]!
+  traitLists: [Variant!]!
 }
 
 input SpeciesCreateInput {
@@ -361,21 +361,12 @@ input TraitFilters {
   speciesId: ID!
 }
 
-type TraitList {
-  enumValueSettings: [EnumValueSetting!]!
-  id: ID!
-  name: String!
-  species: Species!
-  speciesId: ID!
-  traitListEntries: [TraitListEntry!]!
-}
-
 input TraitListCreateInput {
   name: String!
   speciesId: ID!
 }
 
-union TraitListCreateResponse = DuplicateError | InvalidArgumentError | TraitList
+union TraitListCreateResponse = DuplicateError | InvalidArgumentError | Variant
 
 type TraitListEntry {
   defaultDisplayValue: String
@@ -384,7 +375,7 @@ type TraitListEntry {
   required: Boolean!
   trait: Trait!
   traitId: ID!
-  traitList: TraitList!
+  traitList: Variant!
   traitListId: ID!
   valueType: CritterTraitValueType!
 }
@@ -437,6 +428,15 @@ type ValidationConstraint {
 type ValidationError {
   constraints: [ValidationConstraint!]!
   field: String!
+}
+
+type Variant {
+  enumValueSettings: [EnumValueSetting!]!
+  id: ID!
+  name: String!
+  species: Species!
+  speciesId: ID!
+  traitListEntries: [TraitListEntry!]!
 }
 
 union enumValueSettingCreateResponse = DuplicateError | EnumValueSetting | InvalidArgumentError
